Adds configurable android-activity crate
This general Activity glue crate now supersedes the two separate
native-activity and game-activity crates.

The adds the functionality of both crates as modules within this
new crate but provides a single crate + API that downstream can
depend on as an Activity glue layer.

The new crate exposes two (mutually-exclusive) features for
choosing which Activity backend to build:
* "native-activity"
* "game-activity"

This common crate makes it more practical to write code that's
Activity agnostic.