Fixes #141
The `ndk` crate enables `raw-window-handle 0.6` by default (because of https://github.com/rust-mobile/ndk/pull/434#issuecomment-1752089087) which might not be used by consumers of the `android-activity` crate at all, or might (still) be a mismatching version. Even if the `rwh_0x` features are additive, figuring that out leads to cryptic errors and it is best to turn off these defaults completely and leave it to the user to turn it back on in their own `[dependencies]` section if desired.