feat: Add waker() method to AndroidAppWaker
This commit adds an "into_waker()" method to AndroidAppWaker. It converts it
into an `std::task::Waker`, which is the type of waker used by
asynchronous tasks for scheduling. The goal is to allow AndroidAppWaker
to be easily used to set up an asynchronous context.

The implementation is a straightforward wrapper for
`ALooper_acquire/release/wake()`.

Signed-off-by: John Nunley <dev@notgull.net>
Co-authored-by: Robert Bragg <robert@sixbynine.org>