Revert "input: Replace open-coded types with `ndk::event` definitions (#163)"
This reverts commit 51d05d48c8f59a7e9a1d41c1577870842206ce32 for
backwards compatibility with the existing `0.6` releases.

For now, it's creating a lot of busy work having to always make this
revert in order to test various topic branch changes with winit 0.30.

Lets save this breaking change until we have more reasons to break
semver compatibility (in itself this doesn't fix or enable any features,
so we can live without it for now).