Replace unmaintained `android-properties` crate with `ndk` API
I was surprised to find [the `android-properties` crate] with almost 8
*million* downloads; this crate hasn't been updated for almost 5 years
and the repository is a public archive.

Turns out the only published user of this crate is `android-activity`,
so all those downloads are pulled in from folks using this crate (mainly
through `winit`) to build Android apps.

Instead of using an unmaintained, archived crate, use the properties
wrapper and typed API-level wrappers provided in in an upcoming `ndk`
crate release instead:
- https://github.com/rust-mobile/ndk/pull/495
- https://github.com/rust-mobile/ndk/pull/479

[the `android-properties` crate]: https://crates.io/crates/android-properties